DELETE COMPTE

+++ views/comptes/liste.php
@@ -79,14 +79,14 @@
                         <td><?php echo $value['solde'] ?></td>
                         <td>
                             <?php
-                            $date = new DateTime($value['dateCreation']);
-                            echo $date->format('d/m/Y H:s:i');
+                                $date = new DateTime($value['dateCreation']);
+                                echo $date->format('d/m/Y H:s:i');
                             ?>
                         </td>
                         <td><?php echo $value['etat']==1?"<span class='text-success'>Activé</span>":"<span class='text-danger'>Déactivé</span>" ?></td>
                         <!--<td><?php //echo $value['prenom'] ?></td>-->
                         <td>
-                            <a href="compteController/delete/<?= $value["idC"]?>" class="btn btn-x btn-danger"><span class="glyphicon glyphicon-remove"></span></a>
+                            <a href="compteController/delete/<?= $value["idC"]?>" class="btn btn-x btn-danger"><span class="glyphicon glyphicon-remove" onclick="return confirm('Etes-vous sûre de vouloir supprimer?')"></span></a>
                             <a href="compteController/edit/<?= $value["idC"]?>" class="btn btn-x btn-warning"><span class="glyphicon glyphicon-edit"></span></a>
                         </td>
                     </tr>
